Factor some operations out of the `main` binary
This will allow other binaries to use the same methods to merge config
values together and build the database connection string
